components/vertex-components/pyproject.toml
[tool.poetry]
name = "vertex-components"
version = "0.1.0"
authors = ["Example User <user@example.com>"]
description = "KubeFlow components for Vertex AI"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.7",
]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
kfp = ">=2.0.1,<3.0.0"

[tool.poetry.group.dev.dependencies]
google-cloud-aiplatform = "1.24.1"
pytest = ">=7.3.1,<8.0.0"
google-cloud-pipeline-components = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 88
per-file-ignores = [
    "E203",
    "F841"
]
ignore = [
    "E203",
    "F841",
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = "tests"
junit_family = "xunit2"
